% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{abort_dagri}
\alias{abort_dagri}
\title{Abort with a typed dagriculture error}
\usage{
abort_dagri(class, message, details = list())
}
\arguments{
\item{class}{Error class.}

\item{message}{Error message.}

\item{details}{Optional details list.}
}
\description{
Abort with a typed dagriculture error
}
