% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{dagri_ancestors}
\alias{dagri_ancestors}
\title{Get all ancestors}
\usage{
dagri_ancestors(graph, node_id)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}

\item{node_id}{Node ID.}
}
\description{
Get all ancestors
}
