% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{dagri_descendants}
\alias{dagri_descendants}
\title{Get all descendants}
\usage{
dagri_descendants(graph, node_id)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}

\item{node_id}{Node ID.}
}
\description{
Get all descendants
}
