% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{dagri_downstream}
\alias{dagri_downstream}
\title{Get downstream nodes}
\usage{
dagri_downstream(graph, node_id)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}

\item{node_id}{Node ID.}
}
\description{
Get downstream nodes
}
