% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{dagri_edge}
\alias{dagri_edge}
\title{Get an edge from a dagriculture graph}
\usage{
dagri_edge(graph, edge_id)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}

\item{edge_id}{Edge ID.}
}
\description{
Get an edge from a dagriculture graph
}
