% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{dagri_gate}
\alias{dagri_gate}
\title{Get a gate from a dagriculture graph}
\usage{
dagri_gate(graph, id)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}

\item{id}{Gate ID.}
}
\description{
Get a gate from a dagriculture graph
}
