% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{dagri_graph}
\alias{dagri_graph}
\title{Create an empty dagriculture graph}
\usage{
dagri_graph(registry)
}
\arguments{
\item{registry}{A \code{dagri_registry} object.}
}
\description{
Create an empty dagriculture graph
}
