% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_state.R
\name{dagri_recompute_state}
\alias{dagri_recompute_state}
\title{Recompute graph state}
\usage{
dagri_recompute_state(graph)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}
}
\description{
Recompute graph state
}
