% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{dagri_registry}
\alias{dagri_registry}
\title{Define a dagriculture registry}
\usage{
dagri_registry(...)
}
\arguments{
\item{...}{\code{dagri_kind} objects.}
}
\description{
Define a dagriculture registry
}
