% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_editing.R
\name{dagri_remove_gate}
\alias{dagri_remove_gate}
\title{Remove a gate from a dagriculture graph}
\usage{
dagri_remove_gate(graph, id)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}

\item{id}{Gate ID.}
}
\description{
Remove a gate from a dagriculture graph
}
