% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_editing.R
\name{dagri_remove_node}
\alias{dagri_remove_node}
\title{Remove a node from a dagriculture graph}
\usage{
dagri_remove_node(graph, node_id)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}

\item{node_id}{Node ID.}
}
\description{
Remove a node from a dagriculture graph
}
