% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_editing.R
\name{dagri_reopen_gate}
\alias{dagri_reopen_gate}
\title{Reopen a gate in a dagriculture graph}
\usage{
dagri_reopen_gate(graph, id)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}

\item{id}{Gate ID.}
}
\description{
Reopen a gate in a dagriculture graph
}
