% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_editing.R
\name{dagri_resolve_gate}
\alias{dagri_resolve_gate}
\title{Resolve a gate in a dagriculture graph}
\usage{
dagri_resolve_gate(graph, id)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}

\item{id}{Gate ID.}
}
\description{
Resolve a gate in a dagriculture graph
}
