% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{dagri_topo_order}
\alias{dagri_topo_order}
\title{Get topological order}
\usage{
dagri_topo_order(graph, subset = NULL)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}

\item{subset}{Optional subset of nodes.}
}
\description{
Get topological order
}
