% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_editing.R
\name{dagri_update_node}
\alias{dagri_update_node}
\title{Update a node in a dagriculture graph}
\usage{
dagri_update_node(graph, node_id, label = NULL, params = NULL, metadata = NULL)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}

\item{node_id}{Node ID.}

\item{label}{Node label.}

\item{params}{Node parameters.}

\item{metadata}{Node metadata.}
}
\description{
Update a node in a dagriculture graph
}
