% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{dagri_upstream}
\alias{dagri_upstream}
\title{Get upstream nodes}
\usage{
dagri_upstream(graph, node_id)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}

\item{node_id}{Node ID.}
}
\description{
Get upstream nodes
}
