% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagriculture-package.R
\docType{package}
\name{dagriculture-package}
\alias{dagriculture}
\alias{dagriculture-package}
\title{dagriculture: a modern R package starter}
\description{
Replace the package title, description, and example code with project-specific
content before release.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sims1253/dagriculture}
  \item \url{https://sims1253.github.io/dagriculture/}
  \item Report bugs at \url{https://github.com/sims1253/dagriculture/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Scholz \email{dev.scholz@mailbox.org}

}
\keyword{internal}
