% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groots-package.R
\docType{package}
\name{groots-package}
\alias{groots}
\alias{groots-package}
\title{groots: a modern R package starter}
\description{
Replace the package title, description, and example code with project-specific
content before release.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sims1253/groots}
  \item Report bugs at \url{https://github.com/sims1253/groots/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Scholz \email{dev.scholz@mailbox.org}

}
\keyword{internal}
