% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_editing.R
\name{dagriculture_add_edge}
\alias{dagriculture_add_edge}
\title{Add an edge to a dagriculture graph}
\usage{
dagriculture_add_edge(graph, from, to, type = "data", id = NULL, metadata = list())
}
\arguments{
\item{graph}{A \code{dagriculture_graph}.}

\item{from}{Upstream node ID.}

\item{to}{Downstream node ID.}

\item{type}{Edge type.}

\item{id}{Optional Edge ID.}

\item{metadata}{Edge metadata.}
}
\description{
Add an edge to a dagriculture graph
}
