% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_editing.R
\name{dagriculture_add_node}
\alias{dagriculture_add_node}
\title{Add a node to a dagriculture graph}
\usage{
dagriculture_add_node(
  graph,
  id,
  kind,
  label = NULL,
  params = list(),
  metadata = list()
)
}
\arguments{
\item{graph}{A \code{dagriculture_graph}.}

\item{id}{Node ID.}

\item{kind}{Node kind.}

\item{label}{Node label.}

\item{params}{Node parameters.}

\item{metadata}{Node metadata.}
}
\description{
Add a node to a dagriculture graph
}
