% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{dagriculture_ancestors}
\alias{dagriculture_ancestors}
\title{Get all ancestors}
\usage{
dagriculture_ancestors(graph, node_id)
}
\arguments{
\item{graph}{A \code{dagriculture_graph}.}

\item{node_id}{Node ID.}
}
\description{
Get all ancestors
}
