% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_state.R
\name{dagriculture_blocked}
\alias{dagriculture_blocked}
\title{Get blocked nodes}
\usage{
dagriculture_blocked(graph)
}
\arguments{
\item{graph}{A \code{dagriculture_graph}.}
}
\description{
Get blocked nodes
}
