% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{dagriculture_descendants}
\alias{dagriculture_descendants}
\title{Get all descendants}
\usage{
dagriculture_descendants(graph, node_id)
}
\arguments{
\item{graph}{A \code{dagriculture_graph}.}

\item{node_id}{Node ID.}
}
\description{
Get all descendants
}
