% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{dagriculture_downstream}
\alias{dagriculture_downstream}
\title{Get downstream nodes}
\usage{
dagriculture_downstream(graph, node_id)
}
\arguments{
\item{graph}{A \code{dagriculture_graph}.}

\item{node_id}{Node ID.}
}
\description{
Get downstream nodes
}
