% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_state.R
\name{dagriculture_eligible}
\alias{dagriculture_eligible}
\title{Get eligible nodes}
\usage{
dagriculture_eligible(graph)
}
\arguments{
\item{graph}{A \code{dagriculture_graph}.}
}
\description{
Get eligible nodes
}
