% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{dagriculture_gate}
\alias{dagriculture_gate}
\title{Get a gate from a dagriculture graph}
\usage{
dagriculture_gate(graph, id)
}
\arguments{
\item{graph}{A \code{dagriculture_graph}.}

\item{id}{Gate ID.}
}
\description{
Get a gate from a dagriculture graph
}
