% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{dagriculture_graph}
\alias{dagriculture_graph}
\title{Create an empty dagriculture graph}
\usage{
dagriculture_graph(registry)
}
\arguments{
\item{registry}{A \code{dagriculture_registry} object.}
}
\description{
Create an empty dagriculture graph
}
