% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{dagriculture_has_path}
\alias{dagriculture_has_path}
\title{Check path existence}
\usage{
dagriculture_has_path(graph, from, to)
}
\arguments{
\item{graph}{A \code{dagriculture_graph}.}

\item{from}{Source Node ID.}

\item{to}{Target Node ID.}
}
\description{
Check path existence
}
