% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{dagriculture_leaves}
\alias{dagriculture_leaves}
\title{Get graph leaves}
\usage{
dagriculture_leaves(graph)
}
\arguments{
\item{graph}{A \code{dagriculture_graph}.}
}
\description{
Get graph leaves
}
