% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{dagriculture_node}
\alias{dagriculture_node}
\title{Get a node from a dagriculture graph}
\usage{
dagriculture_node(graph, node_id)
}
\arguments{
\item{graph}{A \code{dagriculture_graph}.}

\item{node_id}{Node ID.}
}
\description{
Get a node from a dagriculture graph
}
