% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_state.R
\name{dagriculture_plan}
\alias{dagriculture_plan}
\title{Create an execution plan}
\usage{
dagriculture_plan(graph, targets = NULL)
}
\arguments{
\item{graph}{A \code{dagriculture_graph}.}

\item{targets}{Optional target nodes.}
}
\description{
Create an execution plan
}
