% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_state.R
\name{dagriculture_recompute_state}
\alias{dagriculture_recompute_state}
\title{Recompute graph state}
\usage{
dagriculture_recompute_state(graph)
}
\arguments{
\item{graph}{A \code{dagriculture_graph}.}
}
\description{
Recompute graph state
}
