% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{dagriculture_registry}
\alias{dagriculture_registry}
\title{Define a dagriculture registry}
\usage{
dagriculture_registry(...)
}
\arguments{
\item{...}{\code{dagriculture_kind} objects.}
}
\description{
Define a dagriculture registry
}
