% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_editing.R
\name{dagriculture_remove_node}
\alias{dagriculture_remove_node}
\title{Remove a node from a dagriculture graph}
\usage{
dagriculture_remove_node(graph, node_id)
}
\arguments{
\item{graph}{A \code{dagriculture_graph}.}

\item{node_id}{Node ID.}
}
\description{
Remove a node from a dagriculture graph
}
