% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_editing.R
\name{dagriculture_reopen_gate}
\alias{dagriculture_reopen_gate}
\title{Reopen a gate in a dagriculture graph}
\usage{
dagriculture_reopen_gate(graph, id)
}
\arguments{
\item{graph}{A \code{dagriculture_graph}.}

\item{id}{Gate ID.}
}
\description{
Reopen a gate in a dagriculture graph
}
