% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_editing.R
\name{dagriculture_resolve_gate}
\alias{dagriculture_resolve_gate}
\title{Resolve a gate in a dagriculture graph}
\usage{
dagriculture_resolve_gate(graph, id)
}
\arguments{
\item{graph}{A \code{dagriculture_graph}.}

\item{id}{Gate ID.}
}
\description{
Resolve a gate in a dagriculture graph
}
