% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_state.R
\name{dagriculture_terminal}
\alias{dagriculture_terminal}
\title{Get terminal nodes}
\usage{
dagriculture_terminal(graph)
}
\arguments{
\item{graph}{A \code{dagriculture_graph}.}
}
\description{
Get terminal nodes
}
