% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_editing.R
\name{dagriculture_update_node}
\alias{dagriculture_update_node}
\title{Update a node in a dagriculture graph}
\usage{
dagriculture_update_node(
  graph,
  node_id,
  label = NULL,
  params = NULL,
  metadata = NULL
)
}
\arguments{
\item{graph}{A \code{dagriculture_graph}.}

\item{node_id}{Node ID.}

\item{label}{Node label.}

\item{params}{Node parameters.}

\item{metadata}{Node metadata.}
}
\description{
Update a node in a dagriculture graph
}
