% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_queries.R
\name{dagriculture_upstream}
\alias{dagriculture_upstream}
\title{Get upstream nodes}
\usage{
dagriculture_upstream(graph, node_id)
}
\arguments{
\item{graph}{A \code{dagriculture_graph}.}

\item{node_id}{Node ID.}
}
\description{
Get upstream nodes
}
