% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Return a friendly greeting}
\usage{
hello(name = "world")
}
\arguments{
\item{name}{A name to greet.}
}
\value{
A length-one character vector.
}
\description{
Return a friendly greeting
}
\examples{
hello()
hello("team")
}
