% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_state.R
\name{dagri_plan}
\alias{dagri_plan}
\title{Create an execution plan}
\usage{
dagri_plan(graph, targets = NULL)
}
\arguments{
\item{graph}{A \code{dagri_graph}.}

\item{targets}{Optional target nodes.}
}
\description{
Create an execution plan
}
